Inclusão do NSE.WebApi.Core, para apartar classes em comum entres os serviços
