+++ NerdStoreEnterprise.sln
@@ -13,35 +13,39 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Web", "Web", "{7308C2D4-2D7
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Identity", "Identity", "{E15B8237-02CD-4E17-AD7B-578CFFD7DC85}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NSE.Identity.API", "src\services\NSE.Identity.API\NSE.Identity.API.csproj", "{BD8A66AF-D737-4188-8151-87BEE00BB3C6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NSE.Identity.API", "src\services\NSE.Identity.API\NSE.Identity.API.csproj", "{BD8A66AF-D737-4188-8151-87BEE00BB3C6}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "MVC", "MVC", "{B1D0F94B-0E54-49C7-B9BA-615F94EBD427}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NSE.WebApp.MVC", "src\web\NSE.WebApp.MVC\NSE.WebApp.MVC.csproj", "{9F23B71D-FAB8-4B74-B9CA-153AF9BB6AE2}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NSE.WebApp.MVC", "src\web\NSE.WebApp.MVC\NSE.WebApp.MVC.csproj", "{9F23B71D-FAB8-4B74-B9CA-153AF9BB6AE2}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{70829D1D-2A41-4316-881A-909809560F63}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NSE.Core", "src\building blocks\NSE.Core\NSE.Core.csproj", "{54CCACCF-F492-4D75-8913-10581624F3E8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NSE.Core", "src\building blocks\NSE.Core\NSE.Core.csproj", "{54CCACCF-F492-4D75-8913-10581624F3E8}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Cart", "Cart", "{7EAB9A2E-71F3-4058-82FF-0F6BF4386562}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NSE.Cart.API", "src\services\NSE.Cart.API\NSE.Cart.API.csproj", "{9C516685-1D72-4368-8055-39E070A6FF08}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NSE.Cart.API", "src\services\NSE.Cart.API\NSE.Cart.API.csproj", "{9C516685-1D72-4368-8055-39E070A6FF08}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Catalog", "Catalog", "{10A007ED-61AB-4159-9E62-FCBB62D8A1B0}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NSE.Catalog.API", "src\services\NSE.Catalog.API\NSE.Catalog.API.csproj", "{06D12C5C-0795-4A6B-AB9E-B807F514FEAD}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NSE.Catalog.API", "src\services\NSE.Catalog.API\NSE.Catalog.API.csproj", "{06D12C5C-0795-4A6B-AB9E-B807F514FEAD}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Client", "Client", "{2EF4CE03-3967-4CB7-8AC1-22C8FFDAAC56}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NSE.Client.API", "src\services\NSE.Client.API\NSE.Client.API.csproj", "{39EC8D1B-62E7-425A-8F17-C471A2683811}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NSE.Client.API", "src\services\NSE.Client.API\NSE.Client.API.csproj", "{39EC8D1B-62E7-425A-8F17-C471A2683811}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Payment", "Payment", "{E64FBA5B-3428-45E7-BF5D-BFC4764F7342}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NSE.Payment.API", "src\services\NSE.Payment.API\NSE.Payment.API.csproj", "{A8B9368F-3FCD-480B-A25F-A5689634E10D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NSE.Payment.API", "src\services\NSE.Payment.API\NSE.Payment.API.csproj", "{A8B9368F-3FCD-480B-A25F-A5689634E10D}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Order", "Order", "{6A298574-0EAC-4D4B-823D-BA7226522821}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NSE.Order.API", "src\services\NSE.Order.API\NSE.Order.API.csproj", "{41C4E32C-FF2F-4F48-8426-8AB6E0175601}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NSE.Order.API", "src\services\NSE.Order.API\NSE.Order.API.csproj", "{41C4E32C-FF2F-4F48-8426-8AB6E0175601}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Services", "Services", "{C711B314-8117-4A96-9152-57C52B9685BC}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NSE.WebApi.Core", "src\building blocks\NSE.WebApi.Core\NSE.WebApi.Core.csproj", "{745C1044-A90D-48FD-8A71-BCB6CC773B3C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -81,6 +85,10 @@ Global
 		{41C4E32C-FF2F-4F48-8426-8AB6E0175601}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{41C4E32C-FF2F-4F48-8426-8AB6E0175601}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{41C4E32C-FF2F-4F48-8426-8AB6E0175601}.Release|Any CPU.Build.0 = Release|Any CPU
+		{745C1044-A90D-48FD-8A71-BCB6CC773B3C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{745C1044-A90D-48FD-8A71-BCB6CC773B3C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{745C1044-A90D-48FD-8A71-BCB6CC773B3C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{745C1044-A90D-48FD-8A71-BCB6CC773B3C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -105,6 +113,8 @@ Global
 		{A8B9368F-3FCD-480B-A25F-A5689634E10D} = {E64FBA5B-3428-45E7-BF5D-BFC4764F7342}
 		{6A298574-0EAC-4D4B-823D-BA7226522821} = {B48A9152-EB90-4F8C-A230-3FF68B3D0256}
 		{41C4E32C-FF2F-4F48-8426-8AB6E0175601} = {6A298574-0EAC-4D4B-823D-BA7226522821}
+		{C711B314-8117-4A96-9152-57C52B9685BC} = {AC4F39F2-6A89-4316-9A5E-4CE4AE57812B}
+		{745C1044-A90D-48FD-8A71-BCB6CC773B3C} = {C711B314-8117-4A96-9152-57C52B9685BC}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {B17F97C8-F97C-496A-8D61-6143B9B85C5D}
