+++ src/building blocks/NSE.WebApi.Core/Identity/AppSettings.cs
@@ -1,4 +1,4 @@
-﻿namespace NSE.Identity.API.Extensions
+﻿namespace NSE.WebApi.Core.Identity
 {
     public class AppSettings
     {
