+++ src/services/NSE.Catalog.API/appsettings.Development.json
@@ -7,5 +7,11 @@
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning"
     }
+  },
+  "AppSettings": {
+    "Secret": "MEUSEGREDOSUPERSECRETO",
+    "ExpirationHours": 2,
+    "Issuer": "MeuSistema",
+    "ValidIn": "https://localhost"
   }
 }
