+++ src/services/NSE.Identity.API/Controllers/AuthController.cs
@@ -3,8 +3,8 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
 using NSE.Identity.API.Controllers.Base;
-using NSE.Identity.API.Extensions;
 using NSE.Identity.API.Models;
+using NSE.WebApi.Core.Identity;
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
